factor out and reuse CurveCycleEquipped
